Adds the gitlab provider relation
Adds the gitlab provider relation into the metadata.yaml file,
allowing this charm to create a token for any new gitlab requirer.
When that relation is broken, that token will be removed.

Using gitlab-rails [1] for creating the token, which has the benefit
of not needing to know the user / password in order to generate
a usable token.

The following fields are set in the related gitlab relation data, in the
credentials field:

- host: either the charm application name, or the configured external_url.
  If it's the application name, Juju creates a Kubernetes Service for it,
  (which points towards the right port due to the KubernetesServicePatch)
  which is resolvable and usable internally in the Kubernetes Cluster.
  If an external URL / hostname is used, the expectation that it is resolvable.
- port: always 80. If the host is the application name, the Kubernetes Service
  will be used, which will point towards the right port. If the host is the
  configured external_url, then the related nginx-ingress-integrator charm
  will have already set up an Ingress Route for it.
- api_scheme: https if the config option tls_secret_name is set, http otherwise.
- access-token: The newly generated token for that relation.

[1] https://docs.gitlab.com/ee/user/profile/personal_access_tokens.html#create-a-personal-access-token-programmatically