Uses the Kubernetes Service Patch to patch the Service port
When deploying an application, Juju also creates a service with
the same name as the application. This service has a default port
of 65536/TCP.

Using the charms.observability_libs.v0.kubernetes_service_patch library
allows us to easily patch that Service port to the actual ports being
used by the GitLab Application. In doing so, the related charms will
be able to connect to GitLab through its Kubernetes Service name (
application name) as it is resolvable within the same Kubernetes cluster.

This will not affect the usecase in which an external_url has been configured,
as that is already supposed to be resolvable and routed by NGINX
Ingress.